+++ sdk/resourcemanager/mysql/armmysqlflexibleservers/models.go
@@ -11,6 +11,76 @@ package armmysqlflexibleservers
 
 import "time"
 
+// AdministratorListResult - A List of azure ad administrators.
+type AdministratorListResult struct {
+	// The link used to get the next page of operations.
+	NextLink *string `json:"nextLink,omitempty"`
+
+	// The list of azure ad administrator of a server
+	Value []*AzureADAdministrator `json:"value,omitempty"`
+}
+
+// AdministratorProperties - The properties of an administrator.
+type AdministratorProperties struct {
+	// Type of the sever administrator.
+	AdministratorType *AdministratorType `json:"administratorType,omitempty"`
+
+	// The resource id of the identity used for AAD Authentication.
+	IdentityResourceID *string `json:"identityResourceId,omitempty"`
+
+	// Login name of the server administrator.
+	Login *string `json:"login,omitempty"`
+
+	// SID (object ID) of the server administrator.
+	Sid *string `json:"sid,omitempty"`
+
+	// Tenant ID of the administrator.
+	TenantID *string `json:"tenantId,omitempty"`
+}
+
+// AzureADAdministrator - Represents a Administrator.
+type AzureADAdministrator struct {
+	// The properties of an administrator.
+	Properties *AdministratorProperties `json:"properties,omitempty"`
+
+	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
+	ID *string `json:"id,omitempty" azure:"ro"`
+
+	// READ-ONLY; The name of the resource
+	Name *string `json:"name,omitempty" azure:"ro"`
+
+	// READ-ONLY; The system metadata relating to this resource.
+	SystemData *SystemData `json:"systemData,omitempty" azure:"ro"`
+
+	// READ-ONLY; The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts"
+	Type *string `json:"type,omitempty" azure:"ro"`
+}
+
+// AzureADAdministratorsClientBeginCreateOrUpdateOptions contains the optional parameters for the AzureADAdministratorsClient.BeginCreateOrUpdate
+// method.
+type AzureADAdministratorsClientBeginCreateOrUpdateOptions struct {
+	// Resumes the LRO from the provided token.
+	ResumeToken string
+}
+
+// AzureADAdministratorsClientBeginDeleteOptions contains the optional parameters for the AzureADAdministratorsClient.BeginDelete
+// method.
+type AzureADAdministratorsClientBeginDeleteOptions struct {
+	// Resumes the LRO from the provided token.
+	ResumeToken string
+}
+
+// AzureADAdministratorsClientGetOptions contains the optional parameters for the AzureADAdministratorsClient.Get method.
+type AzureADAdministratorsClientGetOptions struct {
+	// placeholder for future optional parameters
+}
+
+// AzureADAdministratorsClientListByServerOptions contains the optional parameters for the AzureADAdministratorsClient.NewListByServerPager
+// method.
+type AzureADAdministratorsClientListByServerOptions struct {
+	// placeholder for future optional parameters
+}
+
 // Backup - Storage Profile properties of a server
 type Backup struct {
 	// Backup retention days for the server.
@@ -23,6 +93,102 @@ type Backup struct {
 	EarliestRestoreDate *time.Time `json:"earliestRestoreDate,omitempty" azure:"ro"`
 }
 
+// BackupAndExportClientBeginCreateOptions contains the optional parameters for the BackupAndExportClient.BeginCreate method.
+type BackupAndExportClientBeginCreateOptions struct {
+	// Resumes the LRO from the provided token.
+	ResumeToken string
+}
+
+// BackupAndExportClientValidateBackupOptions contains the optional parameters for the BackupAndExportClient.ValidateBackup
+// method.
+type BackupAndExportClientValidateBackupOptions struct {
+	// placeholder for future optional parameters
+}
+
+// BackupAndExportRequest - BackupAndExport API Request
+type BackupAndExportRequest struct {
+	// REQUIRED; Backup Settings
+	BackupSettings *BackupSettings `json:"backupSettings,omitempty"`
+
+	// REQUIRED; Backup Target Store Details
+	TargetDetails BackupStoreDetailsClassification `json:"targetDetails,omitempty"`
+}
+
+// BackupAndExportResponse - Represents BackupAndExport API Response
+type BackupAndExportResponse struct {
+	// End time
+	EndTime *time.Time `json:"endTime,omitempty"`
+
+	// The BackupAndExport operation error response.
+	Error *ErrorResponse `json:"error,omitempty"`
+
+	// Operation progress (0-100).
+	PercentComplete *float64 `json:"percentComplete,omitempty"`
+
+	// The response properties of a backup and export operation.
+	Properties *BackupAndExportResponseProperties `json:"properties,omitempty"`
+
+	// Start time
+	StartTime *time.Time `json:"startTime,omitempty"`
+
+	// The operation status
+	Status *OperationStatus `json:"status,omitempty"`
+
+	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
+	ID *string `json:"id,omitempty" azure:"ro"`
+
+	// READ-ONLY; The name of the resource
+	Name *string `json:"name,omitempty" azure:"ro"`
+
+	// READ-ONLY; The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts"
+	Type *string `json:"type,omitempty" azure:"ro"`
+}
+
+// BackupAndExportResponseProperties - BackupAndExport Response Properties
+type BackupAndExportResponseProperties struct {
+	// Metadata related to backup to be stored for restoring resource in key-value pairs.
+	BackupMetadata *string `json:"backupMetadata,omitempty"`
+
+	// Data transferred in bytes
+	DataTransferredInBytes *int64 `json:"dataTransferredInBytes,omitempty"`
+
+	// Size of datasource in bytes
+	DatasourceSizeInBytes *int64 `json:"datasourceSizeInBytes,omitempty"`
+}
+
+// BackupRequestBase is the base for all backup request.
+type BackupRequestBase struct {
+	// REQUIRED; Backup Settings
+	BackupSettings *BackupSettings `json:"backupSettings,omitempty"`
+}
+
+// BackupSettings - Backup Settings
+type BackupSettings struct {
+	// REQUIRED; Backup Name for the current backup
+	BackupName *string `json:"backupName,omitempty"`
+
+	// Backup Format for the current backup. (CollatedFormat is INTERNAL – DO NOT USE)
+	BackupFormat *BackupFormat `json:"backupFormat,omitempty"`
+}
+
+// BackupStoreDetailsClassification provides polymorphic access to related types.
+// Call the interface's GetBackupStoreDetails() method to access the common type.
+// Use a type switch to determine the concrete type.  The possible types are:
+// - *BackupStoreDetails, *FullBackupStoreDetails
+type BackupStoreDetailsClassification interface {
+	// GetBackupStoreDetails returns the BackupStoreDetails content of the underlying type.
+	GetBackupStoreDetails() *BackupStoreDetails
+}
+
+// BackupStoreDetails - Details about the target where the backup content will be stored.
+type BackupStoreDetails struct {
+	// REQUIRED; Type of the specific object - used for deserializing
+	ObjectType *string `json:"objectType,omitempty"`
+}
+
+// GetBackupStoreDetails implements the BackupStoreDetailsClassification interface for type BackupStoreDetails.
+func (b *BackupStoreDetails) GetBackupStoreDetails() *BackupStoreDetails { return b }
+
 // BackupsClientGetOptions contains the optional parameters for the BackupsClient.Get method.
 type BackupsClientGetOptions struct {
 	// placeholder for future optional parameters
@@ -33,13 +199,9 @@ type BackupsClientListByServerOptions struct {
 	// placeholder for future optional parameters
 }
 
-// CapabilitiesListResult - location capability
-type CapabilitiesListResult struct {
-	// READ-ONLY; Link to retrieve next page of results.
-	NextLink *string `json:"nextLink,omitempty" azure:"ro"`
-
-	// READ-ONLY; A list of supported capabilities.
-	Value []*CapabilityProperties `json:"value,omitempty" azure:"ro"`
+// BackupsClientPutOptions contains the optional parameters for the BackupsClient.Put method.
+type BackupsClientPutOptions struct {
+	// placeholder for future optional parameters
 }
 
 // CapabilityProperties - Location capabilities.
@@ -50,16 +212,19 @@ type CapabilityProperties struct {
 	// READ-ONLY; supported geo backup regions
 	SupportedGeoBackupRegions []*string `json:"supportedGeoBackupRegions,omitempty" azure:"ro"`
 
-	// READ-ONLY; Supported high availability mode
-	SupportedHAMode []*string `json:"supportedHAMode,omitempty" azure:"ro"`
-
-	// READ-ONLY; zone name
-	Zone *string `json:"zone,omitempty" azure:"ro"`
+	// READ-ONLY; A list of supported server versions.
+	SupportedServerVersions []*ServerVersionCapability `json:"supportedServerVersions,omitempty" azure:"ro"`
 }
 
 // CheckNameAvailabilityClientExecuteOptions contains the optional parameters for the CheckNameAvailabilityClient.Execute
 // method.
 type CheckNameAvailabilityClientExecuteOptions struct {
+	// placeholder for future optional parameters
+}
+
+// CheckNameAvailabilityWithoutLocationClientExecuteOptions contains the optional parameters for the CheckNameAvailabilityWithoutLocationClient.Execute
+// method.
+type CheckNameAvailabilityWithoutLocationClientExecuteOptions struct {
 	// placeholder for future optional parameters
 }
 
@@ -107,6 +272,9 @@ type ConfigurationForBatchUpdateProperties struct {
 
 // ConfigurationListForBatchUpdate - A list of server configurations to update.
 type ConfigurationListForBatchUpdate struct {
+	// Whether to reset all server parameters to default.
+	ResetAllToDefault *ResetAllToDefault `json:"resetAllToDefault,omitempty"`
+
 	// The list of server configurations.
 	Value []*ConfigurationForBatchUpdate `json:"value,omitempty"`
 }
@@ -171,19 +339,26 @@ type ConfigurationsClientGetOptions struct {
 // ConfigurationsClientListByServerOptions contains the optional parameters for the ConfigurationsClient.NewListByServerPager
 // method.
 type ConfigurationsClientListByServerOptions struct {
-	// placeholder for future optional parameters
+	// The keyword of the server configuration.
+	Keyword *string
+	// The page of the server configuration.
+	Page *int32
+	// The pageSize of the server configuration.
+	PageSize *int32
+	// The tags of the server configuration.
+	Tags *string
 }
 
 // DataEncryption - The date encryption for cmk.
 type DataEncryption struct {
 	// Geo backup key uri as key vault can't cross region, need cmk in same region as geo backup
-	GeoBackupKeyURI *string `json:"geoBackupKeyUri,omitempty"`
+	GeoBackupKeyURI *string `json:"geoBackupKeyURI,omitempty"`
 
 	// Geo backup user identity resource id as identity can't cross region, need identity in same region as geo backup
 	GeoBackupUserAssignedIdentityID *string `json:"geoBackupUserAssignedIdentityId,omitempty"`
 
 	// Primary key uri
-	PrimaryKeyURI *string `json:"primaryKeyUri,omitempty"`
+	PrimaryKeyURI *string `json:"primaryKeyURI,omitempty"`
 
 	// Primary user identity resource id
 	PrimaryUserAssignedIdentityID *string `json:"primaryUserAssignedIdentityId,omitempty"`
@@ -348,6 +523,22 @@ type FirewallRulesClientListByServerOptions struct {
 	// placeholder for future optional parameters
 }
 
+// FullBackupStoreDetails is used for scenarios where backup data is streamed/copied over to a storage destination.
+type FullBackupStoreDetails struct {
+	// REQUIRED; Type of the specific object - used for deserializing
+	ObjectType *string `json:"objectType,omitempty"`
+
+	// REQUIRED; SASUriList of storage containers where backup data is to be streamed/copied.
+	SasURIList []*string `json:"sasUriList,omitempty"`
+}
+
+// GetBackupStoreDetails implements the BackupStoreDetailsClassification interface for type FullBackupStoreDetails.
+func (f *FullBackupStoreDetails) GetBackupStoreDetails() *BackupStoreDetails {
+	return &BackupStoreDetails{
+		ObjectType: f.ObjectType,
+	}
+}
+
 // GetPrivateDNSZoneSuffixClientExecuteOptions contains the optional parameters for the GetPrivateDNSZoneSuffixClient.Execute
 // method.
 type GetPrivateDNSZoneSuffixClientExecuteOptions struct {
@@ -375,7 +566,7 @@ type HighAvailability struct {
 // Identity - Properties to configure Identity for Bring your Own Keys
 type Identity struct {
 	// Type of managed service identity.
-	Type *string `json:"type,omitempty"`
+	Type *ManagedServiceIdentityType `json:"type,omitempty"`
 
 	// Metadata of user assigned identity.
 	UserAssignedIdentities map[string]any `json:"userAssignedIdentities,omitempty"`
@@ -387,9 +578,59 @@ type Identity struct {
 	TenantID *string `json:"tenantId,omitempty" azure:"ro"`
 }
 
-// LocationBasedCapabilitiesClientListOptions contains the optional parameters for the LocationBasedCapabilitiesClient.NewListPager
+// LocationBasedCapabilitiesClientGetOptions contains the optional parameters for the LocationBasedCapabilitiesClient.Get
 // method.
-type LocationBasedCapabilitiesClientListOptions struct {
+type LocationBasedCapabilitiesClientGetOptions struct {
+	// placeholder for future optional parameters
+}
+
+// LogFile - Represents a logFile.
+type LogFile struct {
+	// The properties of a logFile.
+	Properties *LogFileProperties `json:"properties,omitempty"`
+
+	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
+	ID *string `json:"id,omitempty" azure:"ro"`
+
+	// READ-ONLY; The name of the resource
+	Name *string `json:"name,omitempty" azure:"ro"`
+
+	// READ-ONLY; The system metadata relating to this resource.
+	SystemData *SystemData `json:"systemData,omitempty" azure:"ro"`
+
+	// READ-ONLY; The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts"
+	Type *string `json:"type,omitempty" azure:"ro"`
+}
+
+// LogFileListResult - A List of logFiles.
+type LogFileListResult struct {
+	// The link used to get the next page of operations.
+	NextLink *string `json:"nextLink,omitempty"`
+
+	// The list of logFiles in a server
+	Value []*LogFile `json:"value,omitempty"`
+}
+
+// LogFileProperties - The properties of a logFile.
+type LogFileProperties struct {
+	// Creation timestamp of the log file.
+	CreatedTime *time.Time `json:"createdTime,omitempty"`
+
+	// Last modified timestamp of the log file.
+	LastModifiedTime *time.Time `json:"lastModifiedTime,omitempty"`
+
+	// The size in kb of the logFile.
+	SizeInKB *int64 `json:"sizeInKB,omitempty"`
+
+	// Type of the log file.
+	Type *string `json:"type,omitempty"`
+
+	// The url to download the log file from.
+	URL *string `json:"url,omitempty"`
+}
+
+// LogFilesClientListByServerOptions contains the optional parameters for the LogFilesClient.NewListByServerPager method.
+type LogFilesClientListByServerOptions struct {
 	// placeholder for future optional parameters
 }
 
@@ -529,11 +770,17 @@ type SKUCapability struct {
 	// READ-ONLY; vCore name
 	Name *string `json:"name,omitempty" azure:"ro"`
 
+	// READ-ONLY; Supported high availability mode
+	SupportedHAMode []*string `json:"supportedHAMode,omitempty" azure:"ro"`
+
 	// READ-ONLY; supported IOPS
 	SupportedIops *int64 `json:"supportedIops,omitempty" azure:"ro"`
 
 	// READ-ONLY; supported memory per vCore in MB
 	SupportedMemoryPerVCoreMB *int64 `json:"supportedMemoryPerVCoreMB,omitempty" azure:"ro"`
+
+	// READ-ONLY; Supported zones
+	SupportedZones []*string `json:"supportedZones,omitempty" azure:"ro"`
 
 	// READ-ONLY; supported vCores
 	VCores *int64 `json:"vCores,omitempty" azure:"ro"`
@@ -610,11 +857,17 @@ type ServerBackupProperties struct {
 
 // ServerEditionCapability - Server edition capabilities.
 type ServerEditionCapability struct {
+	// READ-ONLY; Default Sku name
+	DefaultSKU *string `json:"defaultSku,omitempty" azure:"ro"`
+
+	// READ-ONLY; Default storage size
+	DefaultStorageSize *int32 `json:"defaultStorageSize,omitempty" azure:"ro"`
+
 	// READ-ONLY; Server edition name
 	Name *string `json:"name,omitempty" azure:"ro"`
 
-	// READ-ONLY; A list of supported server versions.
-	SupportedServerVersions []*ServerVersionCapability `json:"supportedServerVersions,omitempty" azure:"ro"`
+	// READ-ONLY; A list of supported Skus
+	SupportedSKUs []*SKUCapability `json:"supportedSkus,omitempty" azure:"ro"`
 
 	// READ-ONLY; A list of supported storage editions
 	SupportedStorageEditions []*StorageEditionCapability `json:"supportedStorageEditions,omitempty" azure:"ro"`
@@ -633,6 +886,12 @@ type ServerForUpdate struct {
 
 	// Application-specific metadata in the form of key-value pairs.
 	Tags map[string]*string `json:"tags,omitempty"`
+}
+
+// ServerGtidSetParameter - Server gtid set parameters.
+type ServerGtidSetParameter struct {
+	// The gtid set of server.
+	GtidSet *string `json:"gtidSet,omitempty"`
 }
 
 // ServerListResult - A list of servers.
@@ -721,6 +980,9 @@ type ServerPropertiesForUpdate struct {
 
 	// Storage related properties of a server.
 	Storage *Storage `json:"storage,omitempty"`
+
+	// Server version.
+	Version *ServerVersion `json:"version,omitempty"`
 }
 
 // ServerRestartParameter - Server restart parameters.
@@ -736,9 +998,6 @@ type ServerRestartParameter struct {
 type ServerVersionCapability struct {
 	// READ-ONLY; server version
 	Name *string `json:"name,omitempty" azure:"ro"`
-
-	// READ-ONLY; A list of supported Skus
-	SupportedSKUs []*SKUCapability `json:"supportedSkus,omitempty" azure:"ro"`
 }
 
 // ServersClientBeginCreateOptions contains the optional parameters for the ServersClient.BeginCreate method.
@@ -755,6 +1014,12 @@ type ServersClientBeginDeleteOptions struct {
 
 // ServersClientBeginFailoverOptions contains the optional parameters for the ServersClient.BeginFailover method.
 type ServersClientBeginFailoverOptions struct {
+	// Resumes the LRO from the provided token.
+	ResumeToken string
+}
+
+// ServersClientBeginResetGtidOptions contains the optional parameters for the ServersClient.BeginResetGtid method.
+type ServersClientBeginResetGtidOptions struct {
 	// Resumes the LRO from the provided token.
 	ResumeToken string
 }
@@ -803,6 +1068,9 @@ type ServersClientListOptions struct {
 type Storage struct {
 	// Enable Storage Auto Grow or not.
 	AutoGrow *EnableStatusEnum `json:"autoGrow,omitempty"`
+
+	// Enable IO Auto Scaling or not.
+	AutoIoScaling *EnableStatusEnum `json:"autoIoScaling,omitempty"`
 
 	// Storage IOPS for a server.
 	Iops *int32 `json:"iops,omitempty"`
@@ -881,6 +1149,18 @@ type UserAssignedIdentity struct {
 	PrincipalID *string `json:"principalId,omitempty" azure:"ro"`
 }
 
+// ValidateBackupResponse - Represents ValidateBackup API Response
+type ValidateBackupResponse struct {
+	// The response properties of a pre backup operation.
+	Properties *ValidateBackupResponseProperties `json:"properties,omitempty"`
+}
+
+// ValidateBackupResponseProperties - ValidateBackup Response Properties
+type ValidateBackupResponseProperties struct {
+	// Estimated no of storage containers required for resource data to be backed up.
+	NumberOfContainers *int32 `json:"numberOfContainers,omitempty"`
+}
+
 // VirtualNetworkSubnetUsageParameter - Virtual network subnet usage parameter
 type VirtualNetworkSubnetUsageParameter struct {
 	// Virtual network resource id.
@@ -891,4 +1171,10 @@ type VirtualNetworkSubnetUsageParameter struct {
 type VirtualNetworkSubnetUsageResult struct {
 	// READ-ONLY; A list of delegated subnet usage
 	DelegatedSubnetsUsage []*DelegatedSubnetUsage `json:"delegatedSubnetsUsage,omitempty" azure:"ro"`
+
+	// READ-ONLY; The location name.
+	Location *string `json:"location,omitempty" azure:"ro"`
+
+	// READ-ONLY; The subscription id.
+	SubscriptionID *string `json:"subscriptionId,omitempty" azure:"ro"`
 }