+++ sdk/resourcemanager/mysql/armmysqlflexibleservers/response_types.go
@@ -9,6 +9,36 @@
 
 package armmysqlflexibleservers
 
+// AzureADAdministratorsClientCreateOrUpdateResponse contains the response from method AzureADAdministratorsClient.BeginCreateOrUpdate.
+type AzureADAdministratorsClientCreateOrUpdateResponse struct {
+	AzureADAdministrator
+}
+
+// AzureADAdministratorsClientDeleteResponse contains the response from method AzureADAdministratorsClient.BeginDelete.
+type AzureADAdministratorsClientDeleteResponse struct {
+	// placeholder for future response values
+}
+
+// AzureADAdministratorsClientGetResponse contains the response from method AzureADAdministratorsClient.Get.
+type AzureADAdministratorsClientGetResponse struct {
+	AzureADAdministrator
+}
+
+// AzureADAdministratorsClientListByServerResponse contains the response from method AzureADAdministratorsClient.NewListByServerPager.
+type AzureADAdministratorsClientListByServerResponse struct {
+	AdministratorListResult
+}
+
+// BackupAndExportClientCreateResponse contains the response from method BackupAndExportClient.BeginCreate.
+type BackupAndExportClientCreateResponse struct {
+	BackupAndExportResponse
+}
+
+// BackupAndExportClientValidateBackupResponse contains the response from method BackupAndExportClient.ValidateBackup.
+type BackupAndExportClientValidateBackupResponse struct {
+	ValidateBackupResponse
+}
+
 // BackupsClientGetResponse contains the response from method BackupsClient.Get.
 type BackupsClientGetResponse struct {
 	ServerBackup
@@ -19,8 +49,18 @@ type BackupsClientListByServerResponse struct {
 	ServerBackupListResult
 }
 
+// BackupsClientPutResponse contains the response from method BackupsClient.Put.
+type BackupsClientPutResponse struct {
+	ServerBackup
+}
+
 // CheckNameAvailabilityClientExecuteResponse contains the response from method CheckNameAvailabilityClient.Execute.
 type CheckNameAvailabilityClientExecuteResponse struct {
+	NameAvailability
+}
+
+// CheckNameAvailabilityWithoutLocationClientExecuteResponse contains the response from method CheckNameAvailabilityWithoutLocationClient.Execute.
+type CheckNameAvailabilityWithoutLocationClientExecuteResponse struct {
 	NameAvailability
 }
 
@@ -94,9 +134,14 @@ type GetPrivateDNSZoneSuffixClientExecuteResponse struct {
 	GetPrivateDNSZoneSuffixResponse
 }
 
-// LocationBasedCapabilitiesClientListResponse contains the response from method LocationBasedCapabilitiesClient.NewListPager.
-type LocationBasedCapabilitiesClientListResponse struct {
-	CapabilitiesListResult
+// LocationBasedCapabilitiesClientGetResponse contains the response from method LocationBasedCapabilitiesClient.Get.
+type LocationBasedCapabilitiesClientGetResponse struct {
+	CapabilityProperties
+}
+
+// LogFilesClientListByServerResponse contains the response from method LogFilesClient.NewListByServerPager.
+type LogFilesClientListByServerResponse struct {
+	LogFileListResult
 }
 
 // OperationsClientListResponse contains the response from method OperationsClient.NewListPager.
@@ -137,6 +182,11 @@ type ServersClientListByResourceGroupResponse struct {
 // ServersClientListResponse contains the response from method ServersClient.NewListPager.
 type ServersClientListResponse struct {
 	ServerListResult
+}
+
+// ServersClientResetGtidResponse contains the response from method ServersClient.BeginResetGtid.
+type ServersClientResetGtidResponse struct {
+	// placeholder for future response values
 }
 
 // ServersClientRestartResponse contains the response from method ServersClient.BeginRestart.